better sorting and other bug fixes
Row:
- add move constructor
- add move assignment operator
- fix copy assignment operator
- add column_index method
RowGroup - major rework:
- Rework sorting algorithm
- drop using std::vector container
- add copy and move
Column:
 - fix copy bugs
 DataFrame / example / benchmarks / tests:
 - comply with changes
 - add new sorting algorithm bench

+++ tests/df_tests.h
@@ -131,7 +131,7 @@ TEST(df_sort, dfRowsAscendingSort) {
 
   EXPECT_EQ(sorted_rows.size(), df.row_count());
 
-  long sorted_values[sorted_rows.size()];
+  int sorted_values[sorted_rows.size()];
 
   for (sizetype idx = 0; idx < sorted_rows.size(); idx++) {
     sorted_values[idx] = sorted_rows[idx][col_idx]->value;
